issue: error caused by acc += tl.dot(p.to(v.type.element_ty), v)

+++ flash_attn/flash_attn_triton_amd/fwd_ref.py
@@ -105,7 +105,7 @@ def attention_forward_core_ref_impl(q, k, v, sm_scale, causal, use_exp2):
         print("softmax_lse:", softmax_lse, softmax_lse.shape)
 
     # Compute output
-    o = torch.matmul(softmax, v.to(torch.float32)).to(torch.float16)
+    o = torch.matmul(softmax, v.to(torch.float32))
     if DEBUG_CORE:
         print("o:", o, o.shape)
